# *: optimize error generation, debug log and AllocHandle in the hot write path