--- a/pkg/executor/batch_checker.go
+++ b/pkg/executor/batch_checker.go
@@ -16,7 +16,6 @@ package executor
 
 import (
 	"context"
-	"fmt"
 	"strings"
 
 	"github.com/pingcap/errors"
@@ -213,7 +212,8 @@ func getKeysNeedCheckOneRow(ctx sessionctx.Context, t table.Table, row []types.D
 			}
 			uniqueKeys = append(uniqueKeys, &keyValueWithDupInfo{
 				newKey: key,
-				dupErr: kv.ErrKeyExists.FastGenByArgs(colValStr, fmt.Sprintf("%s.%s", v.TableMeta().Name.String(), v.Meta().Name.String())),
+				dupErr: kv.ErrKeyExists.FastGenByArgs(colValStr,
+					v.TableMeta().Name.String()+"."+v.Meta().Name.String()),
 			})
 		}
 	}