--- a/pkg/session/bench_test.go
+++ b/pkg/session/bench_test.go
@@ -37,6 +37,7 @@ import (
 	"github.com/pingcap/tidb/pkg/util/chunk"
 	"github.com/pingcap/tidb/pkg/util/logutil"
 	"github.com/pingcap/tidb/pkg/util/sqlexec"
+	require "github.com/stretchr/testify/require"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
 )
@@ -1943,4 +1944,66 @@ func TestBenchDaily(t *testing.T) {
 		BenchmarkCompileStmt,
 		BenchmarkAutoIncrement,
 	)
+}
+
+func BenchmarkPipelinedInsertIgnore(b *testing.B) {
+	se, do, st := prepareBenchSession()
+	defer func() {
+		se.Close()
+		do.Close()
+		st.Close()
+	}()
+	mustExecute(se, `create table tmp (id int, dt varchar(512))`)
+	mustExecute(se, `create table src (id int, dt varchar(512))`)
+	for i := 0; i < 100; i++ {
+		mustExecute(se, "begin")
+		for lines := 0; lines < 100; lines++ {
+			mustExecute(se, "insert into src values (42, repeat('x', 512)), (66, repeat('x', 512))")
+		}
+		mustExecute(se, "commit")
+	}
+
+	b.ResetTimer()
+	se.GetSessionVars().BulkDMLEnabled = true
+	se.GetSessionVars().StmtCtx.InInsertStmt = true
+	for i := 0; i < b.N; i++ {
+		se.Execute(context.Background(), "insert ignore into tmp select * from src")
+		warningCount := se.GetSessionVars().StmtCtx.WarningCount()
+		require.Zero(b, warningCount)
+	}
+	b.StopTimer()
+}
+
+func BenchmarkPipelinedInsertOnDuplicate(b *testing.B) {
+	se, do, st := prepareBenchSession()
+	defer func() {
+		se.Close()
+		do.Close()
+		st.Close()
+	}()
+	mustExecute(se, `create table tmp (id int, dt varchar(512), unique key k1(id))`)
+	mustExecute(se, `create table src (id int, dt varchar(512))`)
+	for i := 0; i < 100; i++ {
+		mustExecute(se, "begin")
+		for lines := 0; lines < 100; lines++ {
+			mustExecute(se,
+				fmt.Sprintf(
+					"insert into src values (%d, repeat('x', 512)), (66, repeat('x', 512))",
+					i*100+lines,
+				),
+			)
+		}
+		mustExecute(se, "commit")
+	}
+
+	b.ResetTimer()
+	se.GetSessionVars().BulkDMLEnabled = true
+	se.GetSessionVars().StmtCtx.InInsertStmt = true
+	for i := 0; i < b.N; i++ {
+		se.Execute(context.Background(),
+			"insert into tmp select * from src on duplicate key update dt = values(dt)")
+		warningCount := se.GetSessionVars().StmtCtx.WarningCount()
+		require.Zero(b, warningCount)
+	}
+	b.StopTimer()
 }